Changed to allow the input of command in the strings to be executed and added to the whole string (eg: "The path \\pwd\\ is that." > The path /home/user is that.
Changed the command for getting the text (texto) so that you are able to
use the reserved char to delimit a string inside of said string. As long
as its the only field where you use said reserved chars to delimit a
string since it cuts all the middle text from the first accurrance till the last.